# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Cykloservis API
  version: 0.1.0
  description: Základní veřejné a autentizační endpointy.
servers:
  - url: http://localhost:5001
paths:
...
  /bikes:
    get:
      summary: Výpis kol
      operationId: bikesList
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Seznam kol
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bike'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Vytvoření kola
      operationId: bikesCreate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BikeInput'
      responses:
        '201':
          description: Kolo vytvořeno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bike'
        '400':
          description: Neplatný vstup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '409':
          description: Limit kol dosažen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Příliš velký obrázek
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bikes/deleted:
    get:
      summary: Výpis soft smazaných kol
      operationId: bikesListDeleted
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Seznam kol
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bike'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
  /bikes/{id}:
    get:
      summary: Detail kola
      operationId: bikesGet
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bike'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Aktualizace kola
      operationId: bikesUpdate
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BikeInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bike'
        '400':
          description: Neplatný vstup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '413':
          description: Příliš velký obrázek
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Soft delete kola
      operationId: bikesSoftDelete
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                required: [success]
        '401':
          description: Neautorizováno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bikes/{id}/restore:
    post:
      summary: Obnovení soft smazaného kola
      operationId: bikesRestore
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bike'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bikes/{id}/hard:
    delete:
      summary: Trvalé smazání kola (admin)
      operationId: bikesHardDelete
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bikes/{id}/image:
    post:
      summary: Upload obrázku kola
      operationId: bikesUploadImage
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image: { type: string, format: binary }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bike'
        '400':
          description: Chybný vstup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '413':
          description: Soubor příliš velký
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Bike:
      type: object
      properties:
        id: { type: string, description: 'ID z Mongo ObjectId (server transformuje _id -> id)' }
        title: { type: string }
        type: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        year: { type: integer, minimum: 1900, maximum: 2100 }
        minutesRidden: { type: integer, minimum: 0 }
        imageUrl: { type: string }
        ownerEmail: { type: string, format: email }
        driveBrand: { type: string }
        driveType: { type: string }
        color: { type: string }
        brakes: { type: string }
        suspension: { type: string }
        suspensionType: { type: string }
        specs: { type: string }
        deletedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, ownerEmail]
    BikeInput:
      type: object
      properties:
        title: { type: string }
        type: { type: string }
        manufacturer: { type: string }
        model: { type: string }
        year: { type: integer, minimum: 1900, maximum: 2100 }
        minutesRidden: { type: integer, minimum: 0 }
        imageUrl: { type: string }
        driveBrand: { type: string }
        driveType: { type: string }
        color: { type: string }
        brakes: { type: string }
        suspension: { type: string }
        suspensionType: { type: string }
        specs: { type: string }
      required: [title]
...
    MessageResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
        details: { type: object, additionalProperties: true }
      required: [error, code]
    UnauthorizedResponse:
      type: object
      properties:
        message: { type: string, example: Unauthorized }
      required: [message]
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT